Overhaul of Loader.H (Desc
This is just an upload of the current state (Progress / Work) of the Loader.H, which Loader.H and Loader.C should be working once I've uploaded the files. 

As of now the file reads from an altered version of the original file (removing the windows, Linux, headers for data).